% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addThetaPrior.R
\name{addThetaPrior}
\alias{addThetaPrior}
\title{Add theta prior}
\usage{
addThetaPrior(
  obj,
  dist = "uniform",
  params = list(min = 0, max = 1),
  pname = NULL
)
}
\arguments{
\item{obj}{`CalibSetup` Object}

\item{dist}{string defining distribution e.g., 'normal', 'beta', etc.}

\item{params}{list of parameters e.g., if dist='normal', params is a list with elements 'mean' and 'sd'}

\item{pname}{name of parameter}
}
\value{
An object of class `CalibSetup`
}
\description{
This method adds a prior for the specified theta
}
