% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibSetup.R
\name{CalibSetup}
\alias{CalibSetup}
\title{Pooled Bayesian Model Calibration}
\usage{
CalibSetup(bounds, constraint_func)
}
\arguments{
\item{bounds}{a list with fields of variable names with values of
two dimensional arrays where first element is lower bound
and second element is upper bound.}

\item{constraint_func}{function handle to constraint function on
variables. Defaults to function on checking
bounds}
}
\value{
An object of class `CalibSetup` which is a list with the following
  components:

- `nexp`: number of experiments

- `theta_fixed`: named vector of parameters held fixed during calibration,
  `NULL` until [fixTheta()] is called
}
\description{
This function setups up calibration object
}
