% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefs.R
\name{evalm}
\alias{evalm}
\title{evalm constructor}
\usage{
evalm(obj, ...)
}
\arguments{
\item{obj}{evalm object}

\item{...}{additional arguments passed to method}
}
\value{
An object of class `evalm`
}
\description{
Default constructor for evalm class
}
