% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelBassPca_func.R
\name{ModelBassPca_func}
\alias{ModelBassPca_func}
\title{PCA Based Model Emulator using BASS}
\usage{
ModelBassPca_func(bmod, input_names, exp_ind = NULL, s2 = "MH")
}
\arguments{
\item{bmod}{a object of the type `bassBasis`}

\item{input_names}{cell array of strings of input variable names}

\item{exp_ind}{experiment indices (default: NULL)}

\item{s2}{how to sample error variance (default: 'MH')}
}
\value{
An object of class `ModelBassPca_func`
}
\description{
This function setups up emulator object
}
