% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelmvBayes.R
\name{ModelmvBayes}
\alias{ModelmvBayes}
\title{mvBayes Emulator for Functional Outputs (can use different
BASS/BPPR type emulators)}
\usage{
ModelmvBayes(bmod, input_names, exp_ind = NULL, s2 = "MH", h = FALSE)
}
\arguments{
\item{bmod:}{a object of the type `mvBayes`}

\item{input_names:}{cell array of strings of input variable names}

\item{exp_ind:}{experiment indices (default: NULL)}

\item{s2:}{how to sample error variance (default: 'MH')}

\item{h:}{h representation of warping function (default: FALSE)}
}
\value{
An object of class `ModelmvBayes`
}
\description{
ModelmvBayes Handles larger-dimensional functional responses (e.g., on
large spatial fields) using various inversion tricks. We require any
other covariance e.g., from discrepancy, measurement error, and basis
truncation error) to be diagonal
}
\details{
This function setups up emulator object.
}
