% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelmvBayes_elastic.R
\name{ModelmvBayes_elastic}
\alias{ModelmvBayes_elastic}
\title{mvBayes elastic Emulator for Functional Outputs (can use different
BASS/BPPR type emulators)}
\usage{
ModelmvBayes_elastic(
  bmod,
  bmod_warp,
  input_names,
  exp_ind = NULL,
  s2 = "MH",
  h = FALSE
)
}
\arguments{
\item{bmod}{a object of the type `mvBayes` of aligned functions}

\item{bmod_warp}{a object of the type `mvBayes` of warping functions}

\item{input_names}{cell array of strings of input variable names}

\item{exp_ind}{experiment indices (default: NULL)}

\item{s2}{how to sample error variance (default: 'MH')}

\item{h}{h representation of warping function (default: FALSE)}
}
\value{
An object of class `ModelmvBayes_elastic`
}
\description{
ModelmvBayes_elastic Handles larger-dimensional functional responses (e.g., on
large spatial fields) using various inversion tricks. We require any
other covariance e.g., from discrepancy, measurement error, and basis
truncation error) to be diagonal. Contains a change on the likelihood
}
\details{
This function setups up emulator object.
}
