\name{RImpala-package}
\alias{RImpala-package}
%- Also NEED an '\alias' for EACH other topic documented here.

\title{A Package to interface R and Impala}

\description{
RImpala-package contains the R functions required to connect, execute queries and retrieve back results from Impala. It uses the rJava package to create a JDBC connection to any of the impala servers running on a Hadoop Cluster. 
}

%- maybe also 'usage' for other objects documented here.

\author{
Austin,Sachin,Vikas}

\details{
\tabular{ll}{
Package: \tab RImpala\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2013-09-06\cr
License: \tab file LICENSE\cr
}
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\references{
http://www.cloudera.com/content/cloudera/en/products/cdh/impala.html - Cloudera's page on Impala
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ Impala }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
