% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMCMC.R
\name{setMCMC}
\alias{setMCMC}
\title{Set MCMC Parameters}
\usage{
setMCMC(
  obj,
  nmcmc,
  nburn = 0,
  thin = 1,
  decor = 100,
  start_var_theta = 1e-08,
  start_tau_theta = 0,
  start_var_ls2 = 1e-05,
  start_tau_ls2 = 0,
  start_adapt_iter = 300
)
}
\arguments{
\item{nmcmc}{number of mcmc iterations}

\item{nburn}{number of mcmc burn in iterations (default: 0)}

\item{thin}{number of samples to thin (default: 1)}

\item{decor}{number of mcmc iterations before decorrelation step (default: 100)}

\item{start_var_theta}{start variance of theta proposal (default: 1e-8)}

\item{start_tau_theta}{start tau of theta (default: 0)}

\item{start_var_ls2}{start variance of sigma proposal (default: 1e-5)}

\item{start_tau_ls2}{start tau of sigma (default: 0)}

\item{start_adapt_iter}{number of iterations before adaption
(default: 300)}

\item{`calibPool`}{object}
}
\value{
An object of class `CalibSetup`
}
\description{
This function setups up MCMC parameters for adaptive MCMC,
also includes tempering and decorrelation steps
}
