% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDefs.R
\name{step_m}
\alias{step_m}
\title{Advance a stochastic emulator}
\usage{
step_m(obj, ...)
}
\arguments{
\item{obj}{an emulator model object}

\item{...}{additional arguments passed to the method}
}
\value{
the updated emulator model object
}
\description{
Generic that advances an emulator one MCMC step, used for
  emulators that carry their own posterior samples.
}
