% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{tran_unif}
\alias{tran_unif}
\title{Transform parameters}
\usage{
tran_unif(th, bounds, names)
}
\arguments{
\item{th}{list of of parameters}

\item{bounds}{list of bounds for each parameter that is a two parameter vector with high and low}

\item{names}{vector of variable names}
}
\description{
Transforms variables from the scale 0 to 1 to variable provided by bounds
}
