% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixTheta.R
\name{unfixTheta}
\alias{unfixTheta}
\title{Unfix calibration parameters}
\usage{
unfixTheta(obj, pname = NULL)
}
\arguments{
\item{obj}{`CalibSetup` Object}

\item{pname}{character vector of parameter names to release, or `NULL`
(the default) to release every fixed parameter}
}
\value{
An object of class `CalibSetup`
}
\description{
This method releases parameters fixed by [fixTheta()] so that
  [calibPool()] samples them again.
}
