% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addThetaPrior.R
\name{addJointThetaPrior}
\alias{addJointThetaPrior}
\title{Add joint theta prior}
\usage{
addJointThetaPrior(obj, pnames, log_density_fn)
}
\arguments{
\item{obj}{`CalibSetup` Object}

\item{pnames}{character vector of parameter names to apply joint prior to}

\item{log_density_fn}{function that takes named list of parameter values and returns log-density}
}
\value{
An object of class `CalibSetup`
}
\description{
This method adds a joint prior over multiple theta parameters
}
\examples{
# Joint normal prior on t_1 and t_2
my_joint_prior <- function(params) {
  # params is a list like list(t_1=0.5, t_2=0.3)
  mvtnorm::dmvnorm(c(params$t_1, params$t_2),
                   mean=c(0, 0),
                   sigma=matrix(c(1, 0.5, 0.5, 1), 2, 2),
                   log=TRUE)
}
setup <- addJointThetaPrior(setup, c("t_1", "t_2"), my_joint_prior)

}
