% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVecExperiments.R
\name{addVecExperiments}
\alias{addVecExperiments}
\title{This method adds vector experiments to calibration object}
\usage{
addVecExperiments(
  obj,
  yobs,
  model,
  sd_est,
  s2_df,
  s2_ind,
  meas_error_cor = NULL,
  theta_ind = NULL,
  D = NULL,
  discrep_tau = 1
)
}
\arguments{
\item{obj:}{CalibSetup Object}

\item{yobs:}{a vector of the experiment or observation}

\item{model:}{emulator (currently expecting a object of class
ModelBassPca_func, ModelBassPca_func_mf, or
ModelBpprPca_func}

\item{sd_est:}{estimate of standard deviation}

\item{s2_df:}{degrees of freedom of inverse gamma prior}

\item{s2_ind:}{indices of function}

\item{meas_error_cor:}{measurement error correlation (default: NULL)}

\item{theta_ind:}{indices of theta (default: NULL)}

\item{D:}{discrepency basis (matrix of columns of basis, default:
NaN)}

\item{discrep_tau:}{discrepency sampling tau}
}
\value{
An object of class `CalibSetup`
}
\description{
This method adds vector experiments to calibration object
}
