% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibPool.R
\name{calibPool}
\alias{calibPool}
\title{Pool Calibration}
\usage{
calibPool(setup)
}
\arguments{
\item{setup}{an object of class `CalibSetup`}
}
\value{
a list with the following elements
- theta: mcmc samples of variables
- s2: mcmc samples of error variance
- count: number of counts of acceptance
- count_s2: number of counts of acceptance on error variance
- count_decor: number of times decorrelation occurred
- cov_theta_cand: final theta covariance
- cov_ls2_cand: final error covariance
- pred_curr: current emulator predictions
- discrep_vars: discrepancy coefficients
- llik: log likelihood
- theta_native: mcmc samples of variables in native scale
}
\description{
This function runs pooled Bayesian Model Calibration with adaptive MCMC,
tempering, and decorrelation steps. input theta will be normalized to 0-1 and sampled from uniform priors
}
