% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{cf_bounds}
\alias{cf_bounds}
\title{Compare to bounds}
\usage{
cf_bounds(x, bounds)
}
\arguments{
\item{x}{list of of parameters}

\item{bounds}{list of bounds for each parameter that is a two parameter vector with high and low}
}
\description{
This function compares variables to bounds
}
